admin/src/app/pages/staff-management/create/create.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { CreateComponent } from './create.component';
import { UserService } from '../../../services/user.service';
import { ToastrService } from 'ngx-toastr';
import { FormsModule } from '@angular/forms';
import { of, throwError } from 'rxjs';

describe('CreateComponent', () => {
  let component: CreateComponent;
  let fixture: ComponentFixture<CreateComponent>;
  let userService: jasmine.SpyObj<UserService>;
  let toastService: jasmine.SpyObj<ToastrService>;

  beforeEach(async () => {
    const userServiceSpy = jasmine.createSpyObj('UserService', ['createStaff']);
    const toastServiceSpy = jasmine.createSpyObj('ToastrService', ['success', 'error']);

    await TestBed.configureTestingModule({
      declarations: [CreateComponent],
      imports: [FormsModule],
      providers: [
        { provide: UserService, useValue: userServiceSpy },
        { provide: ToastrService, useValue: toastServiceSpy }
      ]
    }).compileComponents();

    fixture = TestBed.createComponent(CreateComponent);
    component = fixture.componentInstance;
    userService = TestBed.inject(UserService) as jasmine.SpyObj<UserService>;
    toastService = TestBed.inject(ToastrService) as jasmine.SpyObj<ToastrService>;
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render the component template', () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('form')).toBeTruthy();
    expect(compiled.querySelector('input[name="username"]')).toBeTruthy();
    expect(compiled.querySelector('input[name="email"]')).toBeTruthy();
    expect(compiled.querySelector('input[name="firstName"]')).toBeTruthy();
    expect(compiled.querySelector('input[name="lastName"]')).toBeTruthy();
    expect(compiled.querySelector('button[type="submit"]')).toBeTruthy();
  });
  it('should create staff successfully', () => {
    userService.createStaff.and.returnValue(of({}));

    component.onSubmitButton(new MouseEvent('click'));

    expect(toastService.success).toHaveBeenCalledWith('Create staff successfully');
    expect(component.user).toEqual({
      username: '',
      email: '',
      firstName: '',
      lastName: '',
    });
  });

  it('should handle create staff failure', () => {
    const errorResponse = { error: { error: 'Error message' } };
    userService.createStaff.and.returnValue(throwError(errorResponse));

    component.onSubmitButton(new MouseEvent('click'));

    expect(toastService.error).toHaveBeenCalledWith('Create staff failed', 'Error message');
    expect(console.log).toHaveBeenCalledWith(errorResponse);
  });
});